wearableWorker and project reorganization
