+++ Breathe/app/build.gradle
@@ -103,4 +103,15 @@ dependencies {
     // For BLE Scanning
     implementation 'no.nordicsemi.android.support.v18:scanner:1.4.3'
 
+    def work_version = "2.5.0"
+
+    // (Java only)
+    implementation "androidx.work:work-runtime:$work_version"
+
+    // optional - Test helpers -- not used yet
+    androidTestImplementation "androidx.work:work-testing:$work_version"
+
+    // optional - Multiprocess support -- not used yet
+    implementation "androidx.work:work-multiprocess:$work_version"
+
 }
