+++ Breathe/app/src/main/java/com/ybeltagy/breathe/collection/BreatheRepository.java
@@ -1,4 +1,4 @@
-package com.ybeltagy.breathe.persistence;
+package com.ybeltagy.breathe.collection;
 
 import android.app.Application;
