+++ Breathe/app/src/main/java/com/ybeltagy/breathe/ui/BreatheViewModel.java
@@ -7,9 +7,8 @@ import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
 import com.ybeltagy.breathe.data.DiaryEntry;
-import com.ybeltagy.breathe.persistence.BreatheRepository;
+import com.ybeltagy.breathe.collection.BreatheRepository;
 import com.ybeltagy.breathe.data.InhalerUsageEvent;
-import com.ybeltagy.breathe.data.Tag;
 
 import java.time.Instant;
 import java.util.List;
